Change the type of max_signers and min_signers to int32

+++ gateway/src/main.rs
@@ -11,9 +11,9 @@ struct Cli {
     #[clap(subcommand)]
     command: Commands,
     #[arg(long)]
-    max_signers: u8,
+    max_signers: i32,
     #[arg(long)]
-    min_signers: u8,
+    min_signers: i32,
 }
 
 #[derive(Subcommand)]
